requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.2.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.21.6
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pytesseract==0.3.10
Werkzeug==2.2.3
gunicorn==19.7.1
ocr-nanonets-wrapper==1.1